+++ src/main/java/com/huatai/careeragent/task/log/ExecutionLogStatus.java
@@ -4,7 +4,5 @@ public enum ExecutionLogStatus {
     STEP_STARTED,
     STEP_COMPLETED,
     STEP_FAILED,
-    TASK_COMPLETED,
-    @Deprecated SUCCESS,
-    @Deprecated FAILED
+    TASK_COMPLETED
 }